pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "deepflowdb"
version = "0.1.7"
description = "A governance-first reactive platform — bi-temporal event-sourced object store, reactive expressions, and durable workflows backed by PostgreSQL"
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.10"
authors = [
    {name = "Neema Raphael"},
]
keywords = ["reactive", "object-store", "bi-temporal", "event-sourcing", "postgresql", "rls", "deephaven"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Database",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "reaktiv>=0.19.0",
    "psycopg2-binary>=2.9.0",
    "pgserver>=0.1.0; sys_platform != 'linux' or (platform_machine != 'aarch64' and platform_machine != 'arm64')",
    "pixeltable-pgserver>=0.1.0; sys_platform == 'linux' and (platform_machine == 'aarch64' or platform_machine == 'arm64')",
    "dbos>=2.13.0",
    "croniter>=2.0.0",
]

[project.optional-dependencies]
server = [
    "deephaven-server>=0.36.0",
    "deephaven-plugin-ui>=0.30.0",
    "deephaven-plugin-plotly-express>=0.15.0",
    "websockets>=12.0",
    "numpy",
    "pandas",
]
marketdata = [
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "pydantic>=2.0.0",
]
timeseries = [
    "questdb>=2.0.0",
    "httpx>=0.27.0",
]
lakehouse = [
    "pyiceberg[pyarrow]>=0.8.0",
    "duckdb>=1.2.0",
    "pyarrow>=14.0",
    "minio>=7.0",
    "httpx>=0.27.0",
    "s3fs>=2024.0",
    "sqlglot>=26.0",
    "adbc-driver-flightsql>=1.0.0",
]
media = [
    "minio>=7.0",
    "pymupdf>=1.24.0",
    "beautifulsoup4>=4.12.0",
    "pgvector>=0.3.0",
    "httpx>=0.27.0",
]
ai = [
    "google-genai>=1.0",
]
datacube = [
    "perspective-python>=4.0",
    "tornado>=6.0",
    "pyarrow>=14.0",
]
client = [
    "pydeephaven>=0.36.0",
    "pandas",
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]
all = [
    "deepflowdb[server,marketdata,timeseries,lakehouse,media,ai,datacube,client,dev]",
]

[project.urls]
Homepage = "https://github.com/neema2/py-flow"
Repository = "https://github.com/neema2/py-flow"
Issues = "https://github.com/neema2/py-flow/issues"

[tool.hatch.build.targets.wheel]
packages = ["store", "reactive", "bridge", "workflow", "marketdata", "timeseries", "lakehouse", "media", "ai", "streaming", "datacube"]

[tool.ruff]
target-version = "py310"
line-length = 120
src = ["store", "reactive", "bridge", "workflow", "marketdata", "timeseries", "lakehouse", "media", "ai", "streaming", "datacube", "scheduler", "agents", "objectstore", "models"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "UP",     # pyupgrade
    "ANN",    # flake8-annotations (missing type hints)
    "B",      # flake8-bugbear
    "RUF",    # ruff-specific
]
ignore = [
    "ANN101",  # missing self type
    "ANN102",  # missing cls type
    "ANN401",  # Any type allowed
    "E501",    # line length handled by formatter
    "RUF001",  # ambiguous unicode (× etc.) — intentional
    "RUF002",  # ambiguous unicode in docstrings (– etc.) — intentional
    "RUF003",  # ambiguous unicode in comments (σ etc.) — intentional
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["ANN", "B017", "RUF012", "E402"]  # tests: no annotations, broad raises OK, mutable class defaults OK, sys.path OK
"demo_*" = ["ANN", "E402"]   # don't require annotations or top-level imports in demos
"demos/*" = ["ANN", "E402"]
"examples/*" = ["ANN"]
"store/columns/__init__.py" = ["E402"]  # REGISTRY must exist before domain imports
"marketdata/server.py" = ["B008"]       # FastAPI Query() in defaults is standard

[tool.ruff.format]
quote-style = "double"

[tool.ty.environment]
python-version = "3.10"

[tool.pytest.ini_options]
testpaths = ["tests"]
# Disable faulthandler — it conflicts with the JVM's internal SIGSEGV
# handling (safepoint polling). Without this, any DH table derivation
# (last_by, agg_by, sort_descending) crashes the process.
faulthandler_timeout = 0
addopts = "-p no:faulthandler"
